gptel-transient: Add tool selection UI
* gptel-transient.el (gptel--tools-init-value, gptel-menu,
gptel--switch, gptel--switch-category, gptel-tools,
gptel--infix-temperature, gptel--infix-use-tools): Add tool
selection support to `gptel-menu'.  Add new `gptel-tools' command
that directly brings up the sub-menu from where you can select
tools to use.  As with other settings, this selection can be
global, buffer-local or for the next gptel request only.  Change
the keybinding for setting the temperature
(`gptel--infix-temperature') and assign the "-t"
keybinding to toggling tool use instead.

* gptel.el (gptel-use-tools): The tool use toggle now has three
states: t, nil and force.  t is the default behavior -- it will
provide selected tools to the LLM with no further instructions.
Setting it to force will try to force the LLM to use a tool, and
nil turns off tool use.

* gptel-openai.el (gptel--request-data): Force tool use when
`gptel-use-tools' is set to force.

* gptel-gemini.el (gptel--request-data): Force tool use when
`gptel-use-tools' is set to force.

* gptel-anthropic.el (gptel--request-data): Force tool use when
`gptel-use-tools' is set to force.

* gptel-ollama.el (gptel--request-data): I don't know how to force
tool use with Ollama.  Make a note in this function for later.