gptel: Handle pending calls/call results in chat buffers
* gptel.el (gptel--insert-response, gptel--to-string): Handle
pending calls and inserting tool call results in (chat) buffers.
gptel-request's default callbacks delegate this to a new function
`gptel--display-tool-calls'.  Make `gptel--to-string' use
`prin1-to-string' to handle any kind of Lisp object.  This is
required for inserting tool call results in buffers, as they can
be of any type.

(gptel--display-tool-calls): Function to prompt the user with
pending calls in gptel buffers, as well as inserting tool call
results when appropriate.
(gptel--accept-tool-calls, gptel--reject-tool-calls,
gptel--dispatch-tool-calls):  Commands to run or cancel pending
tool calls at the prompt in chat buffers.

* gptel-curl.el (gptel-curl--stream-insert-response): Make the
default streaming callback call the tool-handling UI functions
when required.

* gptel-transient.el (gptel--suffix-send): Make the "echo area"
and "kill ring" callbacks in `gptel--suffix-send' call the
tool-handling UI functions when required.  Tool confirmation is
handled via a minibuffer prompt.  While not ideal, interrupting
the user in the context in which they use echo area/kill ring
redirection is expected to not be too disruptive.