gptel: Make gptel responses front-sticky
* gptel.el (gptel--attach-response-history,
gptel--insert-response, gptel--restore-state): Make the `gptel'
text property front-sticky, so that they are now front-sticky and
rear-nonsticky.  This means text typed at the start of a response
is considered part of the response , while text typed at the end
is not. (#249, #321, #343, #546)

This is another experiment to address the longstanding problem of
being able to edit gptel responses as responses without creating
the many edge cases caused by rear-sticky text properties.  In the
process we hopefully also avoid a whole bunch of API validation
errors caused by whitespace user-edits in the buffer. (#351, #409,

This change is tentative and might be reverted in the future.

* gptel-org.el (gptel-org--restore-state): Concomitant changes.

* gptel-curl.el (gptel-curl--stream-insert-response): Concomitant
changes.