gptel: Make tool arg :type symbols
* gptel.el (gptel-make-tool, gptel--preprocess-tool-args,
gptel--make-tool@preprocess-args): Require argument :type in tool
specifications to be symbols.  Update and improve the
documentation of `gptel-make-tool'.

Internally the :type still needs to be a string so it can be
handled by `json-serialize', so we convert all the :type values to
strings in `gptel--make-tool@preprocess-args' by advising
`gptel--make-tool'.  (There is no cleaner way to modify the
gptel-tool struct constructor that I've been able to find.)

The function `gptel--preprocess-tool-args' was authored by
@ultronozm.