gptel: Set tool-use capability flag for models
Change the symbol to denote the tool-use capability for a model
from 'tool' to 'tool-use'.  Fix errors in capability listings of
models across Anthropic, OpenAI and Gemini.  Note that we do not
test for tool use capability in the UI yet, so the user is free to
send tools with models that don't support it (and deal with the
error).  This will be rectified after implementing a better way to
specify tool-use capability for a model in the package
configuration.

* gptel.el (gptel--openai-models): Fix capabilities.

* gptel-gemini.el (gptel--gemini-models): Fix capabilities.

* gptel-anthropic.el (gptel--anthropic-models): Fix capabilities.