gptel: Breaking change to tool-call internal API
* gptel.el (gptel--format-tool-call, gptel--display-tool-calls,
gptel-request, gptel--handle-tool-use): The tool-call API has been
changed to provide the callback with more information, and to
behave more uniformly across tool call confirmations and tool call
results.  (NOTE: The tool-call internal API remains unstable.)

Suppose CALLBACK is the gptel-request callback.  The new API is as
follows:
- When tool-call(s) require confirmation, CALLBACK is called with
an alist each of whose elements is a list of the tool struct,
argument values and the tool-call function:

 (funcall CALLBACK ((tool args tool-call-callback) ...))

If the CALLBACK wishes to continue the request, it must call the
tool function with args, and call tool-call-callback with the
result.

- When providing tool-call result(s), CALLBACK is called with an
alist each of whose elements is a list of the tool name, argument
values and the result:

 (funcall CALLBACK ((name args result) ...))

The two kinds of calls can be distinguished between by the first
element of each list, which is a tool (struct) in the first case
and the tool name (a string) in the second.

When displaying tool call results in chat buffers, now include the
arguments it was called with.  Add new helper function
`gptel--format-tool-call'.