gptel-curl: Disable ahead-of-time dispatch for parsers
* gptel-curl.el (gptel-curl--parse-response,
gptel-curl--stream-filter, gptel-curl-get-response): Benchmarking
suggests that the ahead-of-time dispatch resolution for
`gptel-curl--parse-stream' and `gptel--parse-response' when using
Curl is not required any more as it does not affect Emacs'
responsiveness enough to be an issue.  Disable this ahead-of-time
dispatch, conditional on further testing.

NOTE: If this change is reverted some parser functions will need
to be rewritten, including the stream parser for gptel-perplexity.