gptel: Change buffer parser property-search method
Parse the buffer using `previous-single-property-change' instead
of `text-property-search-backward'.  This is for compatibility
with overlay-based response bounds tracking.  In the future, we
want to be able to use both overlays and text-properties to track
responses since they can compensate for each other's weaknesses.

Specifically, overlays handle insertions in the middle of response
regions correctly, but are fragile under text deletion.  Text
properties are robust but cannot "swallow" changes to regions
consistently without adding to expensive hooks like
`yank-transform-functions' and `after-change-functions'.

* gptel-openai.el (gptel--parse-buffer): Above described changes.

* gptel-ollama.el (gptel--parse-buffer): Above described changes.

* gptel-gemini.el (gptel--parse-buffer): Above described changes.

* gptel-anthropic.el (gptel--parse-buffer): Above described
changes.