gptel-transient: Make gptel--edit-directive more flexible
* gptel-transient.el (gptel--edit-directive,
gptel--read-with-prefix-help): Convert to a cl-defun with keyword
arguments for the heading, initial text, setup and callback
functions.  The setup function can tweak the buffer state as
appropriate to the (usually Transient suffix) command.  Make
`gptel--read-with-prefix-help' shorter, in preparation for adding
descriptions of more actions.

* gptel-rewrite.el (gptel--infix-rewrite-extra): Allow the rewrite
instruction to be edited in a buffer.  This action is currently
undocumented as it is experimental, and bound to `C-c C-e'. (#623)