gptel-transient: Use `transient--window' for width calculations
* gptel-transient.el: (gptel-system-prompt--format,
gptel--setup-directive-menu, gptel-tools): When discovering the width
of the transient window, use `transient--window' as an argument.
Otherwise, `window-width' will return the width of the active window
where the transient was launched from, and if the frame has been split
horizontally, things get truncated incorrectly.

If `transient-display-buffer-action' has been set to open windows on
the left or right hand side, it also works since transient always fits
the window to the buffer.