gptel-transient: Provide description of send action (#627)
* gptel-transient.el (gptel--describe-infix-context,
gptel--describe-suffix-send, gptel-menu, gptel--suffix-send,
gptel--suffix-context-buffer): In `gptel-menu', provide a verbose
description of what the send command will do.  This is intended to
provide clarity on the effect of the various prompt/response
redirection options gptel provides, as well as to explicitly
indicate the default prompting behavior (which is to send the
selected region or buffer up to point).

Additional changes to `gptel-menu':

- The active context is now indicated in the relevant column
header.
- Incompatible combinations of redirection options are now
disallowed.  This restriction was previously disabled due to a bug
in older versions of Transient.
- Use different faces to indicate active context and selected tools.
- Fix alignment of the "Scope" menu item under Request Parameters.