Explicit tool calls
In order to include tool results in persistent in-buffer context, the gptel
text property has been given new unique values, including (tool . id).

Tool messages needed to be parsed.  This commit covers the OpenAI backend.

They are parsed from their result block, which can be org #+begin_tool_call or
``` for all other modes.  The first line of the block is a read-compatible form
that allows us to rehydrate the call data when sending it to the backend.

To clarify involved implementations, the gptel--display-tool-calls function was
split into itself for calls and gptel--display-tool-results for results.

The response insertion functions were made to respect this and the tool call and
tool results response types were made distinct by consing either 'tool-call or
'tool-result onto them.  See gptel-curl--stream-insert-response and
gptel--insert-response.

Because the org block display in packages like org-modern keeps parameters
visible when folded, the block was chosen.  The truncated call data is visible
to the user.

To prevent org block escape by its contents, the result is escaped for display
and unescaped before sending to the backend.  This takes place in a buffer
containing a clone of the data.

The 'ignore property is introduced and supported by has been de-emphasized.
Users can make use of it to have non-conversation content intermingled with
chat.