bounds overhaul to support more text property types
Property bounds are stored in an alist for compactness.  The alist values are
lists of bounds, which are always lists (trading conses for shorter local
variables) and contain either BEG END or BEG END VAL for text properties like
tool calls that can store an ID.

With work, the tool call ID can be obtained from the buffer data instead of
saved in the local variable as application state, but this is not consistent
across frontends.