gptel: overhaul bounds, support more text property types
Store the chat buffer's gptel bounds in an alist for compactness.
Each entry maps a text designation type (`response', `tool',
`ignore' etc) to a list of bounds.  Each bound is a
list (previously a cons) of the form (BEG END) or (BEG END VAL).
VAL is used for text properties like tool calls that can store an
ID.

The previous variant of gptel's bounds (as a list of conses)
will continue to be supported.

* gptel-org.el (gptel-org--restore-state, gptel-org--save-state):
Update for new bounds data type.

* gptel.el (gptel--in-response-p, gptel--restore-props,
gptel--restore-state): New function 'gptel--restore-props' to
restore the gptel text-property from either of the supported
bounds types.  Other modifications are to handle the fact that
the `gptel' text-property can have many valid values now.