gptel: Improve handling of read only buffers
* gptel.el (gptel-read-only-response-buffer,
gptel--handle-pre-insert): Introduce a new defcustom and don't
hard code "*LLM response*".

* gptel-transient.el (gptel--describe-suffix-send): Inform the user
when the source buffer is read-only and clarify that output will be
directed to the buffer specified in `gptel-read-only-response-buffer`.
Additionally, remove misleading claims about killing content from
a read-only source buffer in "respond in place" mode.

* gptel-transient.el (gptel--suffix-send): Do not attempt to
kill the region in case the buffer is read only. It caused
"kill-region: Buffer is read-only: #<buffer *read-only-test*>"
messages to be emitted, even when it worked correctly.

* gptel-transient.el (gptel--suffix-send): Prevent attempting to kill
the region in read-only buffers, which resulted in "kill-region:
Buffer is read-only: #<buffer *read-only-test*>" messages, despite
functioning correctly.