WIP for comment
The intent here was to make the initial tool result and the results sent from
buffer contents identical.

OpenAI doc says a tool result can be an array, but I got an API error saying it
expected an object, so I guess it must be an array of objects.

Earlier I had removed a string-coercion because it can lose type information: a
string that could be a symbol name and a symbol are indistinguishable after
coercion.  At length, I don't think this matters.  Sending escaped strings as
literals to the model is probably the bigger crime.

Anthropic API might reveal something about the right choice.  Can't hardly read
the Gemini docs.  R1 docs say String only.