gptel-transient: Further improvements to read-only handling
* gptel.el (gptel--read-only-response-buffer,
gptel--handle-pre-insert): Introduce a new variable for fall-back
buffer name, and use it instead of harcoded value.

* gptel-transient.el (gptel--describe-suffix-send): Inform the
user when the whole source buffer, the point, or a part of the
region is read-only, and fall back to the buffer specified
in `gptel--read-only-response-buffer'.

This aims to consider all potential combinations and nuances (such as
stickyness of text properties, or the fact that point and mark can be
swapped) that could influence read-only status, providing users with
meaningful information.

Also, consistently use correct Emacs terminology: "region" instead
of "selection".