Explicit tool calls
Include tool call & result as explicitly recoverable messages within message log
buffers (like chat).  The recovery of tool messages enables re-use of tool
results in several rounds of messages, making RAG-like use cases feasible.  The
:include feature of tools tended to confuse LLMs, resulting in auto-mimicry and
other undesireable behaviors.  These changes corrected this behavior.

To mark explicit tool messages, the gptel text property has been given new
unique values, including (tool . id).  The persistence & rehydration of these
properties is handled in the next commit.

The default callback insertion functions were modified to support a RAW argument
to indicate that they should not propertize the insertion.  This allows re-use
of the several tracking markers and other behavior without treating tool
insertion like response insertion.  See See gptel-curl--stream-insert-response
and gptel--insert-response.

To facilitate creation of a tool message on the backend, the tool call form is
written literally to the beginning of tool messages.  The backend recovers it by
calling read.  After the first read form, the rest of the message is treated
like the tool result.  The result has been coerced to a string through
gptel--to-string before buffer insertion, so the literal buffer string is
treated as the result.  This can erase some type information for backends which
support JSON types besides string (including OpenAI and Gemini).  Further work
may be needed to allow tools to return non-string results to backends which
support them.

This commit only supports the OpenAI backend.  The others will need similar
enhancements:

- unformat and format their tool ids
- match the gptel text property values for (tool . id)
- create tool call and result messages from each recorded call

Buffer encoded tool messages needed to be parsed.  This commit covers the OpenAI
backend.  The strategy is to pcase on the gptel property.  Since tool insertions
have one distinct gptel text property, the call and response message are
recovered at the same time.

This commit also addresses several shortcomings of the UX and display that were
necessarilly tangled due to the path of work.

Because the org block display in packages like org-modern keeps parameters
visible when folded, the block was chosen.  The truncated call data is rendered
as block parameters to easily see which calls were performed while blocks are
folded.  To aid in the recognition of tool messages on text / markdown, the too
message fence has truncated call data appended.  It uses the 'ignore property
even though removing these fences can employ a similar strategy as org mode.

To prevent org block escape by its tool result contents, the tool result is
escaped before insertion for display and unescaped before sending to the
backend.  Unescape takes place in a buffer containing a clone of the data.  The
path for branching context and non-branching context both similarly clone the
buffer to clean it before sending to the backend.

To avoid excessive gptel-separator insertion, new markers were introduced during
insertion, stored in :info as usual.  Using indpendent tool and message markers,
we can infer if we did the last update to the buffer or if another state-path
has done an insertion, meaning we need to re-insert a regular gptel-separator
instead of just a newline or nothing.

The 'ignore property is introduced and supported by has been de-emphasized.
Users can make use of it to have non-conversation content intermingled with
chat.

Fixup: when in org, inform downstream parsing calls to presume org
Fixup: Bug in org unescape
The prev-pt may be ahead of the current point if a read or regex search forward
blows past prev-pt.  In such case, we want to at worst "escape" prev-pt prev-pt.

Fixup: downgrade of malformed tool-use region warning to message
Fixup: propagate local variables into org mode prior to calling backend parsing
Fixup: messed up the switch to rx while trying to make it readable.

Squash: Don't push messages for empty content during parsing (this is later
propagated to other backends)
Squash: Order tool call messages before tool result when parsing chat log

Squash: Tool result string coercion

This commit relates to the conversion of the results of tools.  The result was
to always coerce tool results to strings before handing off to display or to
each backend.

Not all backends support all types of results.  When writing tools, it was
observed that strings are supported by the OpenAI API.  The API docs state that
arrays are also supported, but only an array of objects.  The decision made for
now is to do string coercion of whatever the user passes back.

In the case that the user ever wants to take advantage of non-string types, we
will have to accomodate per-backend.  Backends that support non-string types
will call `read' a second time to recover the tool result.

Earlier I had removed a string-coercion because it can lose type information: A
string that is readable is indistinguishable from an Elisp object after coercing
the object to string.  However, representing strings as escaped strings
in-buffer means we're sending double-escaped results to the backend in many
cases.  The LLMs might not interpret such over-escaping as well.

Anthropic API might reveal something about the right choice.  I Can't navigate
the Gemini docs well enough to judge it clearly.  R1 docs say String only.

Squashed: Update non-streaming callback to feature parity

Result is almost identical to the streaming callback.  The more I looked at each
little behavior, the more I could see that the streaming callback is just better
code and operates in the non-streaming case.

* gptel-anthropic.el:
* gptel-curl.el:
* gptel-openai.el:
* gptel-org.el:
* gptel-transient.el:
* gptel.el: