gptel: Breaking change to gptel-request API
* gptel.el (gptel-request, gptel--handle-tool-use,
gptel--insert-response): Change gptel-request's API for handling
tool use again.  When there are pending tool calls or tool
results, the callback is now called cons cells of the form

 (tool-call   . ((tool args callback) ...))
 (tool-result . ((tool args result) ...))

respectively.  The idea is to support dispatching on more types of
responses (like "reasoning" text blocks) by disambiguating between
them explicitly via symbols like `tool-call' and `tool-result'.
The gptel-request callback can then match against the type of
symbol to define custom behavior.

 (pcase response
   ((pred stringp) ...)                  ;regular response
   (`(,tool-call . call-specs) ...)
   (`(,tool-result . result-specs) ...))

The case of a regular response string remains simple.  See
`gptel--insert-response' for an example of this kind of dispatch.

Update tool handling (in the fsm) for the new API.

* gptel-transient.el (gptel--suffix-send): Update custom callbacks
to the new API.

* gptel-curl.el (gptel-curl--stream-insert-response): Update
`gptel-curl--stream-insert-response' to the new API.