gptel: Add message-marker for fine-tuned spacing
* gptel.el (gptel--insert-response): Avoid excessive
gptel-separator insertion by introducing a message marker in the
insert callback, stored in INFO as usual.  Using an independent
message marker, we can infer if the last update to the buffer was
response text, or if another state-path (like a tool call) has
done an insertion, meaning we need to re-insert a regular
gptel-separator instead of just a newline or nothing.

* gptel-curl.el (gptel-curl--stream-insert-response): Ditto.