gptel: Add support for handling reasoning blocks
Provide options for handling reasoning text ("<think>...</think>")
in the response.  NOTE: This change only covers APIs that include
the reasoning block in the main response.  APIs that provide the
reasoning text in a separate JSON field (such as Deepseek) will be
handled in the following commits.

* gptel.el: (gptel--url-get-response, gptel--insert-response,
gptel-include-reasoning, gptel-request): New user option
`gptel-include-reasoning' to specify how reasoning blocks in the
response should be handled.  They can be omitted from the
response, included normally in the response, included but ignored
in subsequent conversation turns, or redirected to a buffer.
Handle thinking blocks in the process sentinel and the default
non-streaming callback `gptel--insert-response'.

gptel-request's callback is now called with a cons cell of the
form (reasoning . text), where text is the reasoning string.
Mention this in the documentation of gptel-request.

* gptel-curl.el (gptel-curl--sentinel, gptel-curl--stream-filter,
gptel-curl--stream-insert-response): Handle thinking blocks in the
process filter and in the default streaming callback
`gptel-curl--stream-insert-response'.

* gptel-transient.el (transient-format-value, gptel-menu,
gptel--infix-include-reasoning): Add a menu entry to set
`gptel-include-reasoning'.  As with other gptel options, it can be
set globally, for the buffer or for the next request only.