gptel-openai: Fix collision with openrouter parsing (#669)
* gptel-openai.el (gptel-curl--parse-stream): Don't touch the
:reasoning key of fsm-info in OpenAI's `gptel-curl--parse-stream'
method unless using Openrouter.  This was causing collisions with
the parsing of responses from other OpenAI-compatible backends
that include the reasoning text in the main response field.

This whole code path is a kludge, it might be better to just spin
out a gptel-openrouter backend inheriting from gptel-openai.  This
would be similar to how gptel-deepseek works.