gptel: Change how reasoning content is tracked
Stream parsers: Instead of a single :reasoning key in the fsm-info
to hold both the reasoning chunk and the state of the reasoning
block parser (not started/started/ending/done), use two keys
:reasoning and :reasoning-block respectively.  Separating these
functions simplifies the conditionals considerably, and avoids a
limitation where we can't capture both a reasoning chunk and the
end of the reasoning block when they exist in a single filter
run.

* gptel-curl.el (gptel-curl--stream-filter): Use :reasoning and
:reasoning-block instead of overloading :reasoning.

* gptel-openai.el (gptel-curl--parse-stream): Adjust for
new :reasoning-block parameter.

* gptel-openai-extras.el (gptel-curl--parse-stream): Adjust for
new :reasoning-block parameter.

* gptel-anthropic.el (gptel-curl--parse-stream): Adjust for
new :reasoning-block parameter.