gptel-org: Strip property drawers from the prompt
* gptel-org.el: (gptel-org--create-prompt,
gptel-org-ignore-elements, gptel-org--element-end,
gptel-org--strip-elements): Now that prompts in Org mode are
always generated from a temp buffer, we are free to modify the
buffer as we wish.  Add user option `gptel-org-ignore-elements' to
remove specified Org elements from the prompt, and strip property
drawers by default.  (#408, #503)

This behavior can be reverted by setting this
user option to nil, or by selecting a region of text.  (No
filtering is done when selecting a region.)

Eventually prompt filtering will be user-customizable via a hook.

* README.org (Additional Configuration): Mention
`gptel-org-ignore-elements'.