gptel: Create all prompts in temp buffer
Do not load gptel-org.el via `with-eval-after-load', since we
require it anyway when sending commands.  Require gptel when
compiling gptel-org.el.

* gptel.el: (gptel--create-prompt, gptel--wrap-user-prompt-maybe,
gptel-request, gptel--with-buffer-copy, gptel-prompt-filter-hook):
All prompts are now created from a temp buffer.  This includes Org
mode (from before), markdown and strings passed to `gptel-request'.

New macro `gptel--with-buffer-copy' to copy over relevant
local variables and help create prompts from
temp buffers.  This makes it possible to do arbitrary text
transformations to the buffer before sending the request, paving
the way for the new hook `gptel-prompt-filter-hook'.  The name is
tentative.  This hook is intended for transformations like
deleting parts of the buffer the user doesn't want to send, or
replacing shell commands with shell output (#328).

* gptel-org.el (gptel-org--create-prompt): Adjust for
`gptel--with-buffer-copy' and run `gptel-prompt-filter-hook'.

* README (Additional Customization): Mention
`gptel-prompt-filter-hook'.  Document `gptel-post-request-hook'
from before -- this was an omission.