gptel-transient: Don't serialize backend to transient history
* gptel-transient.el (gptel-provider-variable,
transient-format-value, transient-infix-set,
gptel--infix-provider): Saving the backend struct as the value of
`gptel--infix-provider' can serialize it to the Transient history
file.  This serialized value is not always readable, causing
Transient to sometimes throw errors when reading the history file.
Since the history file is shared between all Transients, this can
cause magit (etc) to not function (#693).

`gptel--infix-provider' sets both the model and the backend, but
the backend was the value of the infix.  Invert this to make the
model the value.  The model is a symbol and thus reliably
serializable.  The Transient history file should be much shorter
as a result, as well.