gptel-curl: Handle Deepseek API variance (#709)
* gptel-openai-extras.el (gptel-curl--parse-stream): Some deepseek
model providers will stop supplying "reasoning_content" entries in
the streaming response after the end of the reasoning block.  This
breaks our check in `gptel-deepseek' to detect if we're working
with a reasoning or non-reasoning model.  Fix by explicitly
setting the :reasoning-block key in INFO to indicate that we're in
a reasoning block, then use the value of :reasoning-block to
determine if the model is reasoning-capable.

* gptel-curl.el (gptel-curl--stream-filter):  Set :reasoning-block
when reading reasoning content delivered via a separate JSON
field (i.e. when it's not integrated in the main response stream.)