gptel: Restoring state should not flag the buffer modified (#711)
* gptel.el (gptel--restore-props): Save `buffer-modified-p' before
restoring properties and revert it to its original state.  This
prevents a visited file from being marked as modified until the user
actually makes changes.  This is particularly useful when multiple
files are accessed automatically, such as from the `org-agenda-files'
location (otherwise when exiting Emacs the user is asked about each
file if it needs to be saved).

* gptel-org.el (gptel-org--restore-state): Likewise for Org.