gptel: Don't use generate-new-buffer
When creating temp buffers for rewriting and Org conversion, use
`get-buffer-create' and `generate-new-buffer-name' instead of
`generate-new-buffer'.  The arity of `generate-new-buffer' changed
in Emacs 28, and our call to this with a second arg causes trouble
in Emacs 27.1 We can switch after dropping support for 27.1.

* gptel-org.el (gptel--stream-convert-markdown->org): Make the
above change. (#724)
* gptel-rewrite.el (gptel--suffix-rewrite): Make the above change,
and also use a temp buffer with buffer hooks inhibited to avoid
querying when the buffer is killed.  It is not clear what is
setting up this hook, but in any case nothing of this sort should
run in the gptel-rewrite result accumulation buffer. (#730)