README: Hide authinfo explanation, update Claude 3.7 info
* README.org (Claude 3.7 support, authinfo): Hide authinfo
details.  This is causing more confusion than clarity as readers are
faced with two unfamiliar systems (gptel and authinfo
configuration) instead of one.

Update the Claude 3.7 thinking support text -- using a separate
Anthropic backend for this model is now the official solution, as
no Anthropic-only user options are planned for now to set this.