gptel-org: Fix branching context prompt creation bug
* gptel-org.el (gptel-org--create-prompt): When using branching
context, we always want to include text from (point-min) before
the first headline, as we consider it part of the lineage of the
org element that point is in.  However, if there is a headline
at (point-min), it should only be included if it's part of the
lineage.  This distinction was lost in the recent changes to
branching context handling.  Introduce it again.

* test: Update with tests for the bug fixed in this commit.