gptel: Spacing changes around reasoning, tool result blocks
* gptel.el: (gptel--insert-response, gptel--display-tool-results,
gptel--display-reasoning-stream): Change newline insertion around
tool-result and reasoning blocks.  The new system has the
following behaviors:

- gptel-mode + response-prefixes:  When inserting blocks at the
beginning of a response, account for response prefixes
that end with a newline.

- Consecutive tool-result blocks: Avoid extra newlines between
consecutive tool-result blocks.  This behavior is unchanged from
before.

- `gptel-response-separator' between the end of some response text
and a tool-result block.  This behavior is unchanged from before.

- Newline between a tool-result block and response text that
follows.  Ensure one newline between the end of a tool-result
block and following text.  We want this separator to be
`gptel-response-separator' instead, but the way to enforce this
will require additional tracking markers in
`gptel-curl--stream-insert-response', which makes the happy path
much more complicated to understand.  A better solution is
eventually required here.

- `gptel-response-separator' between the end of a reasoning block
and response text that follows.  We want this whitespace to be
exactly `gptel-response-separator', but with some models,
subsequent text in a stream sometimes begins with extra newlines,
causing a large separation between the reasoning block and
response text.  A better solution is eventually required here as
well.