gptel-openai: Adapt OpenAI tool-ids for Openrouter+Anthropic
* gptel-openai.el (gptel--openai-format-tool-id,
gptel--openai-unformat-tool-id): It is possible to use the
gptel-openai to access Anthropic backends with Openrouter.  In
this case, the tool call ids are not in the call_* format, and
tool calls fail (#747).  Try to work around this by registering
tool call ids as received when they are not in the call_* format.

This will make the chat log not work when switching backends when
both of these conditions are met: (i) tool results are included in
the buffer and (ii) a non-standard tool call id is in effect.  But
there's not much we can do about it since we no longer tell if the
gptel-openai backend in effect wants call_*, toolu_* or some other
format.

* gptel-anthropic.el (gptel--anthropic-unformat-tool-id): Don't
issue a message if the tool call id is not in the expected form.