gptel-anthropic: Handle streaming + thinking + tool-use
* gptel-anthropic.el (gptel-curl--parse-stream): When tool-use is
enabled, we preserve the messages array until the back-and-forth
exchange is complete.  The Anthropic API requires the last
assistant message in this growing messages array (sent back and
forth) to contain a thinking section when thinking and tools are
both enabled.  Ensure that we pack the thinking block correctly
into the messages array (with a signature field) when using both
thinking and tools. (#743)

This fix is not required for the non-streaming case since the
thinking block is already supplied in the form required to send
back to the model.