gptel: Don't use dynamic value of gptel-use-tools (#755)
gptel.el (gptel--tool-use-p, gptel--tool-result-p): Don't depend
on the value of `gptel-use-tools` when handling tool states in the
FSM.  `gptel-use-tools` may be let-bound around a request and be
nil when these checks are called.  Instead, check for `:tools` in
the FSM info, as this is always non-nil if the request was
dispatched with tools.