gptel: Set process coding to utf-8-unix explicitly (#754)
gptel-curl.el (gptel-curl-get-response): Explicitly set the
process coding system to `utf-8-unix' when reading output from
subprocesses.  This ensures correct handling of Unicode text and
prevents encoding-related display issues.  `utf-8-unix' was
previously set explicitly only for Windows platforms. Now do it
for non-windows platforms too.

gptel.el (gptel--url-get-response): Same change, but using
`set-buffer-file-coding-system'.  This has to be set in the
callback to `url-retrieve' and not immediately, as `url-retrieve'
runs `zlib-decompress-region', which requires the buffer to be
unibyte.