gptel-rewrite: Make other modes not clobber rewrite overlay (#744)
When rewriting sections of Lisp or other code where the initial
character is a parenthesis, it results in `show-paren-mode'
highlighting the entire overlay when the point is positioned within
the overlay area.  Other modes like `hl-line-mode' cause similar
issues.  Avoid this by using a very high priority for
`gptel-rewrite-highlight-face'.  (#691)

* gptel-rewrite.el (gptel-rewrite-highlight-face): Inherit from the
`default' face to avoid foreground being the same as
`show-paren-match', which is usually a reverse color.
(gptel--rewrite-callback): Use priority 2000.