gptel: Update tool arg types preprocesing
* gptel.el (gptel--preprocess-tool-args): This (pre-existing)
function is used to convert the :type of tool args from symbols to
strings before running json-serialize.  Update it so that it also
handles the conversion when :type is specified inside a vector,
such as in this example.

  (:name "switch"
   :description "Whether to switch to the new database"
   :allOf [(:type string) (:type string :enum ["true" "false"])]
   :default "false")

Add a TODO item to use the same preprocessor to convert :enum
specifications that are lists instead of arrays.